+++ resources/views/pages/account/profile/security.vue
@@ -5,7 +5,7 @@
             <profile-menu :page="route('profile.security')">
             </profile-menu>
         </template>
-        <div class="my-10 w-full">
+        <div class="my-10 w-full px-3">
             <h1 class="text-xl font-semibold">Update Password</h1>
             <div class="border rounded-lg p-5 my-5 w-full">
                 <h6 class="text-lg my-5">Update Password</h6>
@@ -69,8 +69,11 @@ const deleteAccount=()=>{
         buttons: true,
         dangerMode: true,
     })
-        .then(() => {
-           deleteUser();
+        .then((result) => {
+            if (result.isConfirmed){
+                deleteUser();
+            }
+
         });
 }
